Rename package from farm to pyot
